load OIDC scope and Oauth scope binding config on server startup
