+++ components/org.wso2.carbon.identity.oauth/src/main/java/org/wso2/carbon/identity/oauth2/internal/OAuth2ServiceComponent.java
@@ -18,6 +18,9 @@
 
 package org.wso2.carbon.identity.oauth2.internal;
 
+import org.apache.axiom.om.OMElement;
+import org.apache.axiom.om.impl.builder.StAXOMBuilder;
+import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.osgi.framework.BundleContext;
@@ -35,12 +38,16 @@ import org.wso2.carbon.identity.application.mgt.ApplicationManagementService;
 import org.wso2.carbon.identity.application.mgt.listener.ApplicationMgtListener;
 import org.wso2.carbon.identity.core.util.IdentityCoreInitializedEvent;
 import org.wso2.carbon.identity.event.handler.AbstractEventHandler;
+import org.wso2.carbon.identity.oauth.common.OAuthConstants;
 import org.wso2.carbon.identity.oauth.common.token.bindings.TokenBinderInfo;
 import org.wso2.carbon.identity.oauth.config.OAuthServerConfiguration;
+import org.wso2.carbon.identity.oauth.dto.ScopeDTO;
 import org.wso2.carbon.identity.oauth.internal.OAuthComponentServiceHolder;
 import org.wso2.carbon.identity.oauth2.OAuth2ScopeService;
 import org.wso2.carbon.identity.oauth2.OAuth2Service;
 import org.wso2.carbon.identity.oauth2.OAuth2TokenValidationService;
+import org.wso2.carbon.identity.oauth2.bean.Scope;
+import org.wso2.carbon.identity.oauth2.bean.ScopeBinding;
 import org.wso2.carbon.identity.oauth2.client.authentication.BasicAuthClientAuthenticator;
 import org.wso2.carbon.identity.oauth2.client.authentication.OAuthClientAuthenticator;
 import org.wso2.carbon.identity.oauth2.client.authentication.OAuthClientAuthnService;
@@ -63,7 +70,23 @@ import org.wso2.carbon.identity.user.store.configuration.listener.UserStoreConfi
 import org.wso2.carbon.idp.mgt.IdpManager;
 import org.wso2.carbon.registry.core.service.RegistryService;
 import org.wso2.carbon.stratos.common.listeners.TenantMgtListener;
+import org.wso2.carbon.utils.CarbonUtils;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
+import javax.xml.namespace.QName;
+import javax.xml.stream.XMLInputFactory;
+import javax.xml.stream.XMLStreamException;
+import javax.xml.stream.XMLStreamReader;
+
+import static org.wso2.carbon.identity.oauth2.Oauth2ScopeConstants.PERMISSIONS_BINDING_TYPE;
 import static org.wso2.carbon.identity.oauth2.util.OAuth2Util.checkAudienceEnabled;
 import static org.wso2.carbon.identity.oauth2.util.OAuth2Util.checkIDPIdColumnAvailable;
 
@@ -77,6 +100,13 @@ import static org.wso2.carbon.identity.oauth2.util.OAuth2Util.checkIDPIdColumnAv
 public class OAuth2ServiceComponent {
 
     private static final Log log = LogFactory.getLog(OAuth2ServiceComponent.class);
+    private static final String IDENTITY_PATH = "identity";
+    public static final String NAME = "name";
+    public static final String ID = "id";
+    private static final String DISPLAY_NAME = "displayName";
+    private static final String DESCRIPTION = "description";
+    private static final String PERMISSION = "Permission";
+    private static final String CLAIM = "Claim";
     private BundleContext bundleContext;
 
     @Reference(
@@ -104,6 +134,8 @@ public class OAuth2ServiceComponent {
     protected void activate(ComponentContext context) {
 
         try {
+            loadScopeConfigFile();
+            loadOauthScopeBinding();
             int tenantId = PrivilegedCarbonContext.getThreadLocalCarbonContext().getTenantId();
             boolean isRecordExist = OAuthTokenPersistenceFactory.getInstance().getScopeClaimMappingDAO().
                     hasScopesPopulated(tenantId);
@@ -460,5 +492,149 @@ public class OAuth2ServiceComponent {
     protected void unsetIdpManager(IdpManager idpManager) {
 
         OAuth2ServiceComponentHolder.getInstance().setIdpManager(null);
+    }
+
+    private static void loadScopeConfigFile() {
+
+        List<ScopeDTO> listOIDCScopesClaims = new ArrayList<>();
+        String configDirPath = CarbonUtils.getCarbonConfigDirPath();
+        String confXml =
+                Paths.get(configDirPath, IDENTITY_PATH, OAuthConstants.OIDC_SCOPE_CONFIG_PATH).toString();
+        File configFile = new File(confXml);
+        if (!configFile.exists()) {
+            log.warn("OIDC scope-claim Configuration File is not present at: " + confXml);
+            return;
+        }
+
+        XMLStreamReader parser = null;
+        try (InputStream stream = new FileInputStream(configFile)) {
+
+            parser = XMLInputFactory.newInstance().createXMLStreamReader(stream);
+            StAXOMBuilder builder = new StAXOMBuilder(parser);
+            OMElement documentElement = builder.getDocumentElement();
+            Iterator iterator = documentElement.getChildElements();
+            while (iterator.hasNext()) {
+                ScopeDTO scope = new ScopeDTO();
+                OMElement omElement = (OMElement) iterator.next();
+                String configType = omElement.getAttributeValue(new QName(ID));
+                scope.setName(configType);
+
+                String displayName = omElement.getAttributeValue(new QName(DISPLAY_NAME));
+                if (StringUtils.isNotEmpty(displayName)) {
+                    scope.setDisplayName(displayName);
+                } else {
+                    scope.setDisplayName(configType);
+                }
+
+                String description = omElement.getAttributeValue(new QName(DESCRIPTION));
+                if (StringUtils.isNotEmpty(description)) {
+                    scope.setDescription(description);
+                }
+
+                scope.setClaim(loadClaimConfig(omElement));
+                listOIDCScopesClaims.add(scope);
+            }
+        } catch (XMLStreamException e) {
+            log.warn("Error while streaming OIDC scope config.", e);
+        } catch (IOException e) {
+            log.warn("Error while loading OIDC scope config.", e);
+        } finally {
+            try {
+                if (parser != null) {
+                    parser.close();
+                }
+            } catch (XMLStreamException e) {
+                log.error("Error while closing XML stream", e);
+            }
+        }
+        OAuth2ServiceComponentHolder.getInstance().setOIDCScopesClaims(listOIDCScopesClaims);
+    }
+
+    private static String[] loadClaimConfig(OMElement configElement) {
+
+        StringBuilder claimConfig = new StringBuilder();
+        Iterator it = configElement.getChildElements();
+        while (it.hasNext()) {
+            OMElement element = (OMElement) it.next();
+            if (CLAIM.equals(element.getLocalName())) {
+                String commaSeparatedClaimNames = element.getText();
+                if (StringUtils.isNotBlank(commaSeparatedClaimNames)) {
+                    claimConfig.append(commaSeparatedClaimNames.trim());
+                }
+            }
+        }
+
+        String[] claim;
+        if (claimConfig.length() > 0) {
+            claim = claimConfig.toString().split(",");
+        } else {
+            claim = new String[0];
+        }
+        return claim;
+    }
+
+    private static void loadOauthScopeBinding() {
+
+        List<Scope> scopes = new ArrayList<>();
+        String configDirPath = CarbonUtils.getCarbonConfigDirPath();
+        String confXml = Paths.get(configDirPath, IDENTITY_PATH, OAuthConstants.OAUTH_SCOPE_BINDING_PATH).toString();
+        File configFile = new File(confXml);
+        if (!configFile.exists()) {
+            log.warn("OAuth scope binding File is not present at: " + confXml);
+            return;
+        }
+
+        XMLStreamReader parser = null;
+        try (InputStream stream = new FileInputStream(configFile)) {
+
+            parser = XMLInputFactory.newInstance()
+                    .createXMLStreamReader(stream);
+            StAXOMBuilder builder = new StAXOMBuilder(parser);
+            OMElement documentElement = builder.getDocumentElement();
+            Iterator iterator = documentElement.getChildElements();
+            while (iterator.hasNext()) {
+                OMElement omElement = (OMElement) iterator.next();
+                String scopeName = omElement.getAttributeValue(new QName(NAME));
+                String displayName = omElement.getAttributeValue(new QName(DISPLAY_NAME));
+                String description = omElement.getAttributeValue(new QName(DESCRIPTION));
+                List<String> bindingPermissions = loadScopePermissions(omElement);
+                ScopeBinding scopeBinding = new ScopeBinding(PERMISSIONS_BINDING_TYPE, bindingPermissions);
+                List<ScopeBinding> scopeBindings = new ArrayList<>();
+                scopeBindings.add(scopeBinding);
+                Scope scope = new Scope(scopeName, displayName, scopeBindings, description);
+                scopes.add(scope);
+            }
+        } catch (XMLStreamException e) {
+            log.warn("Error while streaming oauth-scope-bindings config.", e);
+        } catch (IOException e) {
+            log.warn("Error while loading oauth-scope-bindings config.", e);
+        } finally {
+            try {
+                if (parser != null) {
+                    parser.close();
+                }
+            } catch (XMLStreamException e) {
+                log.error("Error while closing XML stream", e);
+            }
+        }
+        OAuth2ServiceComponentHolder.getInstance().setOauthScopeBinding(scopes);
+    }
+
+    private static List<String> loadScopePermissions(OMElement configElement) {
+
+        List<String> permissions = new ArrayList<>();
+        Iterator it = configElement.getChildElements();
+        while (it.hasNext()) {
+            OMElement element = (OMElement) it.next();
+            Iterator permissionIterator = element.getChildElements();
+            while (permissionIterator.hasNext()) {
+                OMElement permissionElement = (OMElement) permissionIterator.next();
+                if (PERMISSION.equals(permissionElement.getLocalName())) {
+                    String permission = permissionElement.getText();
+                    permissions.add(permission);
+                }
+            }
+        }
+        return permissions;
     }
 }
